Extract the endpoint user id lookup into EndpointUser
Minimal API handlers have no ControllerBase.User, so every authenticated
endpoint repeats the same walk from IHttpContextAccessor to the claims
principal. Roughly forty endpoints are still to be migrated, so this puts
the walk in one place before the pattern spreads.
